+++ ecms_base/modules/custom/ecms_blocks/src/Plugin/Block/SiteNotificationsBlock.php
@@ -11,6 +11,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Cache\Cache;
 
 /**
  * Provides a listing of promos referenced by the node.
@@ -117,6 +118,14 @@ class SiteNotificationsBlock extends BlockBase implements ContainerFactoryPlugin
    */
   protected function blockAccess(AccountInterface $account): AccessResult {
     return AccessResult::allowedIfHasPermission($account, 'access content');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getCacheTags(): array {
+    // We want to rebuild when notifications change.
+    return Cache::mergeTags(parent::getCacheTags(), ['node_list:notification']);
   }
 
 }